flexboard/backend/requirements.txt
# ------------------------------- #
# Section 1, Lecture 3 | Go down to see what YOU downloaded
# ------------------------------- #
# Exact Version of Packages used:
#
# atomicwrites==1.4.0
# attrs==20.3.0
# bcrypt==3.2.0
# certifi==2020.12.5
# cffi==1.14.5
# chardet==4.0.0
# click==7.1.2
# colorama==0.4.4
# coverage==5.5
# dnspython==2.1.0
# ecdsa==0.14.1
# email-validator==1.1.2
# fastapi==0.63.0
# greenlet==1.0.0
# h11==0.12.0
# idna==2.10
# iniconfig==1.1.1
# Jinja2==2.11.3
# MarkupSafe==1.1.1
# packaging==20.9
# passlib==1.7.4
# pluggy==0.13.1
# psycopg2==2.8.6
# py==1.10.0
# pyasn1==0.4.8
# pycparser==2.20
# pydantic==1.8.1
# pyparsing==2.4.7
# pytest==6.2.3
# pytest-cov==2.11.1
# python-dotenv==0.17.0
# python-jose==3.2.0
# python-multipart==0.0.5
# requests==2.25.1
# rsa==4.7.2
# six==1.15.0
# SQLAlchemy==1.4.11
# starlette==0.13.6
# toml==0.10.2
# typing-extensions==3.7.4.3
# urllib3==1.26.4
# uvicorn==0.13.4
# ------------------------------- #


# Mine :)

# -------- #
# Endpoint #
# -------- #
# for api endpoint + server
fastapi
uvicorn==0.13.4
python-multipart==0.0.5  # this is for form data. OAuth2PasswordRequestForm.


# -------- #
# Env Vars #
# -------- #
# for env variables     | for keeping environment variables in a file ( .env )
python-dotenv



# -------- #
# Database #
# -------- #
# for databases         | ORM, we will not write raw SQL
sqlalchemy

# adapter               | REQUIRED for sqlalchemy, SQLAlchemy depends on psycopg2 or other database drivers to communicate with the database!
psycopg2



# -------- #
# Pydantic #
# -------- #
# for validation        | email validation, etc.
pydantic[email]



# --------------------- #
# Bcrypt, Security, JWT #
# --------------------- #
# for salting strings        | bcrypt
passlib[bcrypt]==1.7.4

# JWT
python-jose==3.2.0



# ------- #
# Testing #
# ------- #
pytest
requests

# -------- #
# Coverage #
# -------- #
pytest-cov==2.11.1